All Components working properly

+++ frontend/src/components/OrdersScreen.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
 import { listOrders, deleteOrder } from '../actions/orderActions';
@@ -20,9 +20,22 @@ function OrdersScreen(props) {
   }, [successDelete]);
 
   const deleteHandler = (order) => {
-    dispatch(deleteOrder(order._id));
+    if(loadingDelete)
+    {
+      return "Loading...";
+    }
+    else if(errorDelete)
+    {
+      return error;
+    }
+    else if(window.confirm("Are you sure?"))
+    {
+      dispatch(deleteOrder(order._id));
+    }
   }
-  return loading ? <div>Loading...</div> :
+
+  return loading ? <div className="fa fa-spinner fa-3x fa-spin primary">Loading...</div> :
+    error ? <div>{error}</div> :
     <div className="content content-margined">
 
       <div className="order-header">
@@ -47,13 +60,13 @@ function OrdersScreen(props) {
           <tbody>
             {orders.map(order => (<tr key={order._id}>
               <td>{order._id}</td>
-              <td>{order.createdAt}</td>
-              <td>{order.totalPrice}</td>
+              <td>{order.createdAt.substring(0,10)}</td>
+              <td>{order.totalPrice.toFixed(2)}</td>
               <td>{order.user.name}</td>
-              <td>{order.isPaid.toString()}</td>
-              <td>{order.paidAt}</td>
-              <td>{order.isDelivered.toString()}</td>
-              <td>{order.deliveredAt}</td>
+              <td>{order.isPaid ? "Yes" : "No" }</td>
+              <td>{order.isPaid ? order.paidAt.substring(0, 10) : "Not known" }</td>
+              <td>{order.isDelivered ? "Yes" : "No" }</td>
+              <td>{order.deliveredAt ? order.deliveredAt : "Not done" }</td>
               <td>
                 <Link to={"/order/" + order._id} className="button secondary" >Details</Link>
                 {' '}
